+++ Once with a Colonist/Assets/Script/Modules/Gameplay/Character/PlayerCamera.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace TendedTarsier
+namespace TendedTarsier.Script.Modules.Gameplay.Character
 {
     public class PlayerCamera : MonoBehaviour
     {
