+++ Once with a Colonist/Assets/Script/Modules/Gameplay/Configs/InventoryConfig.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using TendedTarsier.Script.Modules.Gameplay.Field;
 using TendedTarsier.Script.Modules.Gameplay.Services.Inventory;
 using TendedTarsier.Script.Modules.Gameplay.Services.Inventory.Items;
 using UnityEngine;
