+++ Once with a Colonist/Assets/Script/Modules/Gameplay/Field/SimpleItem.cs
@@ -1,4 +1,4 @@
-namespace TendedTarsier
+namespace TendedTarsier.Script.Modules.Gameplay.Field
 {
     public class SimpleItem : MapItemBase
     {
