+++ Once with a Colonist/Assets/Script/Modules/Gameplay/Services/Inventory/Items/PerformTest.cs
@@ -25,7 +25,6 @@ namespace TendedTarsier.Script.Modules.Gameplay.Services.Inventory.Items
 
             if (_tilemapService.GetTile(tilemap, targetPosition) == TileModel.TileType.Stone)
             {
-                Debug.Log($"{nameof(Tilemap)} already has {TileModel.TileType.Stone} at {targetPosition}");
                 return false;
             }
