+++ Once with a Colonist/Assets/Script/Modules/Gameplay/Services/Tilemaps/TileModel.cs
@@ -2,7 +2,7 @@ using System;
 using UnityEngine;
 using UnityEngine.Tilemaps;
 
-namespace TendedTarsier
+namespace TendedTarsier.Script.Modules.Gameplay.Services.Tilemaps
 {
     [Serializable]
     public class TileModel
