+++ Once with a Colonist/Assets/Script/Modules/Gameplay/Services/Tilemaps/TilesControllers/PropsAltar.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using UnityEngine;
 
-namespace TendedTarsier
+namespace TendedTarsier.Script.Modules.Gameplay.Services.Tilemaps.TilesControllers
 {
     public class PropsAltar : MonoBehaviour
     {
