+++ Once with a Colonist/Assets/Script/Utilities/MemoryPack/FormatterProviders/ReactiveCollectionFormatter.cs
@@ -2,7 +2,7 @@
 using MemoryPack;
 using UniRx;
 
-namespace TendedTarsier
+namespace TendedTarsier.Script.Utilities.MemoryPack.FormatterProviders
 {
     public class ReactiveCollectionFormatter<TElement> : MemoryPackFormatter<ReactiveCollection<TElement>>
     {
